server/db_builder/requirements.txt
formic-py3==0.1.0
mutagen==1.40.0